pyproject.toml
[project]
name = "scribepy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.3",
    "openai-whisper>=20250625",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
]
